requirements.txt
tensorflow
numpy
matplotlib
python-dotenv